test: Add platform resolution tests from build-push-action
Extracted platform resolution logic to a separate module for better
testability and added comprehensive tests to ensure platform resolution
works correctly for both user-supplied and host-derived platforms.

Changes:
- Created platform-utils.ts with resolveRemoteBuilderPlatforms function
- Added comprehensive tests for platform resolution
- Tests cover user-supplied platforms and fallback to host architecture
- Restructured code for better separation of concerns
